Add CLI as an example
Having a CLI to quickly test changes to the library makes development a
lot easier. After considering creating a workspace for the library and a
CLI binary, the decision was made to use an example instead. The binary
will never be used outside the scope of the library, so having a
separate crate feels like overkill.