Derive simple constructors
While porting structs to the unified packet format, and changing them to
respect Rust's API guidelines, it becomes necessary to create
constructors for them. These are all basic, and simply construct the
struct from the arguments. Luckily, such simple constructors can be
derived so that we don't have to create them by hand.