Implement generic car status packet
A generic implementation of the car status packet has been added that
can be used with either F1 2018 or F1 2019. Based on Rust's API
guidelines, a common set of traits has been derived or implemented for
the new types.

One challenge with this packet has been the tyre model, which has
changed significantly between F1 2018 and F1 2019. Further testing is
necessary to evaluate if the proposed solution works.