Bring API spec and packet type back into header
The packet headers sent by the F1 games contains the version of the
game, also referred to as the packet format or API specification, and
the packet type. Both are only necessary to decode the packet, and have
therefore been dropped from the header that is published to clients of
the library. But for decoding, it is helpful to have easy access to
them, so they have been brought back.